Script to curl the message shown on the top on the Summit at snowqualimi website(https://summitatsnoqualmie.com') and send a SMS when it changes.
Just created this script as I wanted to get  notified of the updates on the website in real time.
You can modify this script to scrap any website to extract the info you are looking for by updating line 25 for the site url and line 40 for the data extraction.

I redacted my personal Twilio accountSid and authtoken and obviously my personal mobile number. 
Twilio provides a free number to send SMS on singup, you can use the number and auth details as it shows on your profile after you singup.